fix warning in CPlayerOperations::Stop()

+++ xbmc/interfaces/json-rpc/PlayerOperations.cpp
@@ -262,7 +262,7 @@ JSONRPC_STATUS CPlayerOperations::Stop(const CStdString &method, ITransportLayer
   {
     case Video:
     case Audio:
-      CApplicationMessenger::Get().MediaStop(true, parameterObject["playerid"].asInteger());
+      CApplicationMessenger::Get().MediaStop(true, (int)parameterObject["playerid"].asInteger());
       return ACK;
 
     case Picture:
